requirements.txt
streamlit==1.31.0
torch==2.1.2
torchvision==0.16.2
transformers==4.37.2
Pillow==10.2.0
gtts==2.5.0
easyocr==1.7.1
numpy==1.26.3
opencv-python-headless==4.9.0.80
SpeechRecognition==3.10.0
pyaudio==0.2.14